Add files via upload
This project obtains datasets, cleans them, and was able to derive some exciting insights from them. The project was able to get the highest grossing companies, highest grossing directors, and highest grossing movies. It also went further to try to find a correlation between gross and other items in the datasets. We concluded that budget has the highest correlation with gross with 0.75 closely followed by votes. We can confidently recommend that gross incomes increase with the budget. The higher the budget, the higher the gross revenue.